requirements.txt
torch
torchvision
torchaudio
numpy
python-chess
zstandard
tqdm